# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: GTW IOT API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
securityDefinitions:
  UserSecurity:
    type: apiKey
    in: header
    name: x-api-key
security:
  - UserSecurity: []
paths:
...
  /user/createUser:
    x-swagger-router-controller: userManage
    post:
      tags:
      - user
      summary: create a new user
      description: create a new user
      # used as the method name of the controller
      operationId: createUser
      parameters:
      - in: body
        name: body
        description: User object thats need to be added
        required: true
        schema:
          $ref: '#/definitions/UserModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/UserModel"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/updateUser:
    x-swagger-router-controller: userManage
    put:
      tags:
      - user
      summary: update a user
      description: update a user
      # used as the method name of the controller
      operationId: updateUser
      parameters:
      - in: body
        name: body
        description: User object thats need to be updated
        required: true
        schema:
          $ref: '#/definitions/UserModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /user/getSettings:
    x-swagger-router-controller: userManage
    get:
      tags:
      - user
      summary: get all user settings
      description: get all user gettings
      # used as the method name of the controller
      operationId: getSettings
      parameters:
      - name: userID
        in: query
        description: get all user settings
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/UserSettings"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/updateSettings:
    x-swagger-router-controller: userManage
    post:
      tags:
      - user
      summary: update all user settings
      description: update all user gettings
      # used as the method name of the controller
      operationId: updateSettings
      parameters:
      - in: body
        name: body
        description: User object thats need to be updated
        required: true
        schema:
          $ref: '#/definitions/UserModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  UserModel:
    type: object
    properties:
      UserID:
        type: string
      Active:
        type: number
      EmailAddress:
        type: string
      Password:
        type: string
...
  UserSettings:
    type: object
    properties:
      SettingName:
        type: string
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string